Minor fixes

+++ stablehlo/reference/Element.cpp
@@ -364,10 +364,8 @@ Element min(const Element &e1, const Element &e2) {
 
 Element rsqrt(const Element &el) {
   return mapWithUpcastToDouble(
-      el, [](double e) { return static_cast<double>(1) / std::sqrt(e); },
-      [](std::complex<double> e) {
-        return static_cast<double>(1) / std::sqrt(e);
-      });
+      el, [](double e) { return 1.0 / std::sqrt(e); },
+      [](std::complex<double> e) { return 1.0 / std::sqrt(e); });
 }
 
 Element sine(const Element &el) {
